refactor: code splitting on skeleton loaders
