+++ src/components/country-list.js
@@ -3,7 +3,7 @@ import { nothing } from 'lit-html';
 
 import '@appnest/masonry-layout/masonry-layout';
 import './country-card';
-import './country-card-loader';
+import './skeleton-loaders/country-card-skeleton';
 import './scroll-to-top-button';
 
 import fontStyles from '../styles/font-styles';
@@ -149,31 +149,32 @@ export default class CountryList extends LitElement {
           if (i === this.loadedItems - 1) {
             return html`
               <div class="loader" id="loadMoreSentinel">
-                <country-card-loader></country-card-loader>
+                <country-card-skeleton></country-card-skeleton>
               </div>
             `;
           }
 
-          if (i > this.loadedItems - 1) {
-            return html` <country-card-loader></country-card-loader> `;
-          }
+          return html` <country-card-skeleton></country-card-skeleton> `;
+          // if (i > this.loadedItems - 1) {
+          //   return html` <country-card-skeleton></country-card-skeleton> `;
+          // }
 
-          return html`
-            <div class="country-card-item">
-              <country-card
-                activeFilter="${this.count}"
-                cases="${e.cases.total}"
-                country="${e.country.replace(/-/g, ' ')}"
-                coverage="${this.coverage}"
-                deaths="${e.deaths.total}"
-                newCases="${e.cases.new}"
-                newDeaths="${e.deaths.new}"
-                rank="${e.position}"
-                recovered="${e.cases.recovered}"
-                time="${e.time}"
-              ></country-card>
-            </div>
-          `;
+          // return html`
+          //   <div class="country-card-item">
+          //     <country-card
+          //       activeFilter="${this.count}"
+          //       cases="${e.cases.total}"
+          //       country="${e.country.replace(/-/g, ' ')}"
+          //       coverage="${this.coverage}"
+          //       deaths="${e.deaths.total}"
+          //       newCases="${e.cases.new}"
+          //       newDeaths="${e.deaths.new}"
+          //       rank="${e.position}"
+          //       recovered="${e.cases.recovered}"
+          //       time="${e.time}"
+          //     ></country-card>
+          //   </div>
+          // `;
         });
     }
